+++ src/main/java/minkostplan/application/UIcontroller/HomepageController.java
@@ -1,12 +1,6 @@
 package minkostplan.application.UIcontroller;
-
-import minkostplan.application.entity.Users;
-import minkostplan.application.usecase.UserUtil;
-import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class HomepageController {
@@ -29,23 +23,7 @@ public class HomepageController {
 
     @GetMapping("/admin")
     public String admin(){
-        return "adminPage";
+        return "adminpage";
     }
 
-
-    @GetMapping("/profile")
-    public String profilepage(Authentication authentication, Model model){
-        Users user = UserUtil.getCurrentUser();
-        System.out.println("User" + user);
-        model.addAttribute("User",user);
-        // Your POST method logic here
-        return"profilePage";
-    }
-
-
-    @PostMapping("/changeProfilePicture")
-    public String profilepicturechange(){
-
-        return"profilePage";
-    }
 }
